+++ app/src/main/java/com/example/amandaungco/buynothingmatcher/model/User.java
@@ -8,7 +8,7 @@ import org.json.JSONObject;
 public class User {
     private String location;
     private String name;
-    private int userId;
+    private Long userId;
     private String email;
 
     public String getEmail() {
@@ -27,13 +27,11 @@ public class User {
         this.name = name;
     }
 
-    public int getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
-        this.userId = userId;
-    }
+    public void setUserId(Long userId) { this.userId = userId; }
 
     public String getLocation() {
         return location;
@@ -46,12 +44,10 @@ public class User {
     public static User fromJson(JSONObject userJSONData) throws JSONException { //throws json
         User newUser = new User();
 
-
         newUser.setName(userJSONData.getString("name"));
         newUser.setLocation(userJSONData.getString("location"));
-        newUser.setUserId(userJSONData.getInt("id"));
+        newUser.setUserId(userJSONData.getLong("id"));
         newUser.setEmail(userJSONData.getString("email"));
-
 
         return newUser;
